Update PolarisWallet.md
Hi @keeganquigley 
Thank you for your feedback. We’ve made the following updates:
Vesting Percentage: We’ve adjusted our submission to meet the 50% DOT vesting requirement. The earlier mention of 30% was a mistake, and it’s now fixed.
Milestones 3-5 Formatting: We’ve corrected the formatting for Milestones 3-5 to match the template guidelines.
Budget Adjustment: We’ve adjusted our budget requirements
Case Studies: We’ve shared a document with case studies of our previous projects to showcase our experience and capabilities.